+++ fdb-record-layer-core/src/main/java/com/apple/foundationdb/record/provider/foundationdb/FDBExceptions.java
@@ -20,6 +20,7 @@
 
 package com.apple.foundationdb.record.provider.foundationdb;
 
+import com.apple.foundationdb.FDBError;
 import com.apple.foundationdb.FDBException;
 import com.apple.foundationdb.annotation.API;
 import com.apple.foundationdb.record.RecordCoreException;
@@ -168,18 +169,18 @@ public class FDBExceptions {
         }
         if (ex instanceof FDBException) {
             FDBException fdbex = (FDBException)ex;
-            switch (fdbex.getCode()) {
-                case 1007:    // transaction_too_old
+            switch (FDBError.fromCode(fdbex.getCode())) {
+                case TRANSACTION_TOO_OLD:
                     return new FDBStoreTransactionIsTooOldException(fdbex).addLogInfo(logInfo);
-                case 1020:    // not_committed
+                case NOT_COMMITTED:
                     return new FDBStoreTransactionConflictException(fdbex).addLogInfo(logInfo);
-                case 1031: // transaction_timed_out
+                case TRANSACTION_TIMED_OUT:
                     return new FDBStoreTransactionTimeoutException(fdbex).addLogInfo(logInfo);
-                case 2101:    // transaction_too_large
+                case TRANSACTION_TOO_LARGE:
                     return new FDBStoreTransactionSizeException(fdbex).addLogInfo(logInfo);
-                case 2102:    // key_too_large
+                case KEY_TOO_LARGE:
                     return new FDBStoreKeySizeException(fdbex).addLogInfo(logInfo);
-                case 2103:    // value_too_large
+                case VALUE_TOO_LARGE:
                     return new FDBStoreValueSizeException(fdbex).addLogInfo(logInfo);
                 default:
                     if (fdbex.isRetryable()) {